5 min. Sieve of Eratosthenes. Refer video.

+++ cpp/baekjoon-algorithm-basic/012-300-1929-소수구하기-Math.cpp
@@ -5,24 +5,24 @@ using namespace std;
 
 int main()
 {
-    int m, n;
+    long m, n;
     cin >> m >> n;
 
     vector<bool> C(n, false);
-    vector<int> P;
+    vector<long> P;
 
-    for (int i = 2; i <= n; i++)
+    for (long i = 2; i <= n; i++)
     {
         if (!C[i])
         {
             if (i >= m)
                 P.push_back(i);
-            for (int j = i*i; j <= n; j+=i)
+            for (long j = i*i; j <= n; j+=i)
                 C[j] = true;
         }
     }
 
-    for(int p : P)
+    for(long p : P)
     {
         cout << p << '\n';
     }
